backend/requirements.txt
fastapi
uvicorn
scikit-learn
requests
google-generativeai
python-dotenv
